refactor(meshtastic): update imports to align with meshtastic library changes
Replace references to `meshtastic.mesh_interface.BROADCAST_NUM` and `facade.meshtastic.ble_interface.BLEInterface` with direct imports from the `meshtastic` package. This cleans up the dependency access and ensures compatibility with the current library structure.
